challenge 1/requirements.txt
azure-quantum==0.18.2109.162713